Move node Stringer implementation to separate file
